refactor(api): extrai validators áudio/speech + misc apikey de validation.ts
Decomposição god-file (Tier 1 / #3501), 4ª fatia. Move os validators de áudio/
speech + misc API-key para módulo co-localizado, byte-idêntico ao inline original.

- providers/validation/audioMiscProviders.ts — deepgram/assemblyai/elevenlabs/
  inworld/kie/aws-polly/bailian/reka/maritalk/nlpcloud/runwayml/nous-research/poe

validation.ts: 2333 → 1771 (−562; acumulado −2751 das 4 fatias). Imports órfãos
removidos do host: runway/maritalk (config), awsSigV4, e buildRekaHeaders/
buildKeyHeaders/buildTokenHeaders do headers.

Validação: typecheck:core, eslint (0/0), check:cycles, check:file-size (módulo
<800), 151 testes de validação existentes + 2 de caracterização — todos verdes.
